requirements.txt

# Core Data Processing
polars>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
scikit-learn>=1.4.0

# Deep Learning & Optimization
torch>=2.2.0
optuna>=3.5.0
einops>=0.7.0
xgboost>=2.0.0

# Asynchronous & Network
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0

# Visualization & Dashboard
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
streamlit>=1.30.0

# Financial Analysis
pandas_ta>=0.3.14b0
binance-connector>=3.0.0
PyWavelets>=1.5.0

# Utilities
tqdm>=4.66.1
pyyaml>=6.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
nbformat>=5.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
